Crud Completed and Report Pendent

+++ database/migrations/2021_05_03_085730_create_dqc841_table.php
@@ -32,7 +32,7 @@ class CreateDqc841Table extends Migration
                 ->on($this->tableDqc84)
                 ->onDelete('CASCADE');
             $table->string('parts_no', 15)->unique()->index('parts_no');
-            $table->integer('unt_usg');
+            $table->decimal('unt_usg', 10, 6);
             $table->longText('description');
             $table->longText('ref_designator')->nullable();
             $table->timestamps();
